refactor unittests into separate files for each plugin
